requirements.txt
aws-cdk-lib==2.54.0
bandit==1.7.4
boto3==1.26.32
cdk-nag==2.21.35
constructs>=10.0.0,<11.0.0
flake8==5.0.4
pytest==6.2.5
